1. Additional RateTopic() and DeleteRate() logic wich throw and hadles the exceptions 2. Small ICrude<> changes: Parameters now have a default value so no need to set it, if you are using standart querry parameters 3. Adding DeleteRateModel and RateTopicModel in models 4. Adding BaseController in Controller folder, which derrived from the ControllerBase and adding new convinient way to use fluent validation. 5. TopicController was derrived from BaseController and changed accordingly.

+++ WebApp.BusinessLogic/Services/TopicService.cs
@@ -13,7 +13,7 @@ public class TopicService : ITopicService
     private readonly IUnitOfWork unitOfWork;
     private readonly IMapper mapper;
 
-    private double CalculateActivity(ICollection<MessageModel> messages)
+    private static double CalculateActivity(ICollection<MessageModel> messages)
     {
         double decayConstant = 0.1; // Adjust based on desired decay rate
 
@@ -30,14 +30,20 @@ public class TopicService : ITopicService
         this.mapper = mapper ?? throw new ArgumentNullException(nameof(mapper));
     }
 
-    public async Task<IEnumerable<TopicSummaryModel>> GetAllAsync(TopicQueryParametersModel queryParameters)
+    public async Task<IEnumerable<TopicSummaryModel>> GetAllAsync(TopicQueryParametersModel? queryParameters = null)
     {
+        if (queryParameters == null)
+        {
+            var allTopics = await this.unitOfWork.TopicRepository.GetAllAsync();
+            return allTopics.Select(m => this.mapper.Map<TopicSummaryModel>(m));
+        }
+
         IEnumerable<Topic> query = new List<Topic>();
 
         // Start with the query for all topics
         if (queryParameters.Size > 0)
         {
-            query = await this.unitOfWork.TopicRepository.GetRangeAsync(queryParameters.Page, queryParameters.Size);
+            query = await this.unitOfWork.TopicRepository.GetRangeAsync(queryParameters.Page - 1, queryParameters.Size);
         }
 
         // Apply filtering if a search term is provided
@@ -64,6 +70,20 @@ public class TopicService : ITopicService
         // Execute the query and project results to TopicDto
         var topics = query.Select(t => this.mapper.MapWithExceptionHandling<TopicSummaryModel>(t));
 
+        foreach (var topic in query)
+        {
+            if (topic.Messages.Count == 0 || topic.Messages is null)
+            {
+                topics.First(t => t.Id == topic.Id).ActivityScore = 0;
+            }
+            else
+            {
+                var topicMessages = topic.Messages.Select(m => this.mapper.MapWithExceptionHandling<MessageModel>(m));
+                topics.First(t => t.Id == topic.Id).ActivityScore =
+                    CalculateActivity((ICollection<MessageModel>)topicMessages);
+            }
+        }
+
         return topics;
     }
 
@@ -79,7 +99,7 @@ public class TopicService : ITopicService
         else
         {
             var topicMessages = topicEntity.Messages.Select(m => this.mapper.MapWithExceptionHandling<MessageModel>(m));
-            topicModel.ActivityScore = this.CalculateActivity((ICollection<MessageModel>)topicMessages);
+            topicModel.ActivityScore = CalculateActivity((ICollection<MessageModel>)topicMessages);
         }
 
         return topicModel;
@@ -121,16 +141,33 @@ public class TopicService : ITopicService
 
     public async Task RateTopic(int userId, int topicId, int stars)
     {
-        await this.unitOfWork.TopicStarsRepository.AddAsync(new TopicStars
+        try
         {
-            UserId = userId, TopicId = topicId, StarCount = stars,
-        });
+            // Check if the user has already rated this topic
+            var existingRating = await this.unitOfWork.TopicStarsRepository.GetByIdAsync(userId, topicId);
+            existingRating.StarCount = stars;
+            this.unitOfWork.TopicStarsRepository.Update(existingRating);
+        }
+        catch (NullReferenceException)
+        {
+            await this.unitOfWork.TopicStarsRepository.AddAsync(new TopicStars
+            {
+                UserId = userId, TopicId = topicId, StarCount = stars,
+            });
+        }
 
         await this.unitOfWork.SaveAsync();
     }
 
     public async Task RemoveRate(int userId, int topicId)
     {
+        var stars = await this.unitOfWork.TopicStarsRepository.GetByIdAsync(userId, topicId);
+
+        if (stars is null)
+        {
+            throw new InvalidOperationException("Stars with such id for this topic does not exist");
+        }
+
         await this.unitOfWork.TopicStarsRepository.DeleteByIdAsync(userId, topicId);
         await this.unitOfWork.SaveAsync();
     }
